pyproject.toml
[project]
name = "shoe-cosmeticks"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduard",email = "barned85@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "mypy (>=1.16.1,<2.0.0)",
    "django (>=5.2.4,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "environs (>=14.2.0,<15.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "django-phonenumber-field (>=8.1.0,<9.0.0)",
    "eventlet (>=0.40.1,<0.41.0)",
    "coverage (>=7.9.2,<8.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-image-cropping (>=1.7,<2.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119